Add files via upload
Description:
This repository contains JavaScript code demonstrating the use of constructor functions and object instantiation. The main focus is on the Student constructor, creating instances of the Student class, and calling the info method to display student information.

File Structure:

index.html: A basic HTML file for testing and displaying results in the browser.
script.js: The JavaScript file containing the Student constructor and instance creation.                                                                                                                                                       How to Use:

Clone the repository: git clone <repository_url>
Open index.html in a browser to see the results in the console.